refactor!: change package naming
from the official Java tutorial: "Package names
are written in all lower case..."
