--- app/src/main/java/io/github/ClownGraphics/Rerenderer/App.java
@@ -1,14 +1,0 @@
-/*
- * This source file was generated by the Gradle 'init' task
- */
-package io.github.ClownGraphics.Rerenderer;
-
-public class App {
-    public String getGreeting() {
-        return "Hello World!";
-    }
-
-    public static void main(String[] args) {
-        System.out.println(new App().getGreeting());
-    }
-}
