--- app/src/main/java/io/github/ClownGraphics/Rerenderer/color/MonotoneTexture.java
@@ -1,29 +1,0 @@
-package io.github.ClownGraphics.Rerenderer.color;
-
-
-import io.github.ClownGraphics.Rerenderer.math.Barycentric;
-import javafx.scene.paint.Color;
-import java.util.Objects;
-
-public class MonotoneTexture implements Texture {
-    private final ColorRGB color;
-
-    public MonotoneTexture(final ColorRGB color) {
-        Objects.requireNonNull(color);
-
-        this.color = color;
-    }
-
-    public MonotoneTexture(final Color color) {
-        Objects.requireNonNull(color);
-
-        this.color = new ColorRGB(color);
-    }
-
-    @Override
-    public ColorRGB get(final Barycentric b) {
-        Objects.requireNonNull(b);
-
-        return color;
-    }
-}
