--- app/src/main/java/io/github/ClownGraphics/Rerenderer/color/Texture.java
@@ -1,9 +1,0 @@
-package io.github.ClownGraphics.Rerenderer.color;
-
-
-import io.github.ClownGraphics.Rerenderer.math.Barycentric;
-
-public interface Texture {
-    public ColorRGB get(final Barycentric b);
-}
-
