--- app/src/main/java/io/github/ClownGraphics/Rerenderer/math/affine_transform/Translation.java
@@ -1,4 +1,0 @@
-package io.github.ClownGraphics.Rerenderer.math.affine_transform;
-
-public class Translation {
-}
