--- app/src/main/java/io/github/ClownGraphics/Rerenderer/model/Camera.java
@@ -1,4 +1,0 @@
-package io.github.ClownGraphics.Rerenderer.model;
-
-public class Camera {
-}
