--- app/src/main/java/io/github/ClownGraphics/Rerenderer/model/Model.java
@@ -1,4 +1,0 @@
-package io.github.ClownGraphics.Rerenderer.model;
-
-public class Model {
-}
