--- app/src/main/java/io/github/ClownGraphics/Rerenderer/model/Object.java
@@ -1,4 +1,0 @@
-package io.github.ClownGraphics.Rerenderer.model;
-
-public interface Object {
-}
