--- app/src/main/java/io/github/ClownGraphics/Rerenderer/triangulation/TriangulationException.java
@@ -1,9 +1,0 @@
-package io.github.ClownGraphics.Rerenderer.triangulation;
-
-
-public class TriangulationException extends RuntimeException {
-
-    public TriangulationException(String errorMessage) {
-        super(errorMessage);
-    }
-}
