--- app/src/main/java/io/github/ClownGraphics/Rerenderer/triangulation/Vector2f.java
@@ -1,6 +1,0 @@
-package io.github.ClownGraphics.Rerenderer.triangulation;
-
-public interface Vector2f {
-    float x();
-    float y();
-}
