+++ app/src/main/java/io/github/clowngraphics/rerenderer/App.java
@@ -1,0 +1,14 @@
+/*
+ * This source file was generated by the Gradle 'init' task
+ */
+package io.github.clowngraphics.rerenderer;
+
+public class App {
+    public String getGreeting() {
+        return "Hello World!";
+    }
+
+    public static void main(String[] args) {
+        System.out.println(new App().getGreeting());
+    }
+}
