+++ app/src/main/java/io/github/clowngraphics/rerenderer/color/Texture.java
@@ -1,0 +1,7 @@
+package io.github.clowngraphics.rerenderer.color;
+
+import io.github.clowngraphics.rerenderer.math.Barycentric;
+
+public interface Texture {
+    public ColorRGB get(final Barycentric b);
+}
