+++ app/src/main/java/io/github/clowngraphics/rerenderer/math/affine_transform/Scale.java
@@ -1,0 +1,4 @@
+package io.github.clowngraphics.rerenderer.math.affine_transform;
+
+public class Scale {
+}
