+++ app/src/main/java/io/github/clowngraphics/rerenderer/model/Model.java
@@ -1,0 +1,4 @@
+package io.github.clowngraphics.rerenderer.model;
+
+public class Model {
+}
