+++ app/src/main/java/io/github/clowngraphics/rerenderer/model/Object.java
@@ -1,0 +1,4 @@
+package io.github.clowngraphics.rerenderer.model;
+
+public interface Object {
+}
