+++ app/src/main/java/io/github/clowngraphics/rerenderer/triangulation/TriangulationException.java
@@ -1,0 +1,8 @@
+package io.github.clowngraphics.rerenderer.triangulation;
+
+public class TriangulationException extends RuntimeException {
+
+    public TriangulationException(String errorMessage) {
+        super(errorMessage);
+    }
+}
