+++ app/src/main/java/io/github/clowngraphics/rerenderer/triangulation/Vector2f.java
@@ -1,0 +1,7 @@
+package io.github.clowngraphics.rerenderer.triangulation;
+
+public interface Vector2f {
+    float x();
+
+    float y();
+}
