--- app/src/test/java/io/github/ClownGraphics/Rerenderer/triangulation/ReadOnlyVector2f.java
@@ -1,19 +1,0 @@
-package io.github.ClownGraphics.Rerenderer.triangulation;
-
-public class ReadOnlyVector2f implements Vector2f {
-    private final float x;
-    private final float y;
-
-    public ReadOnlyVector2f(float x, float y) {
-        this.x = x;
-        this.y = y;
-    }
-
-    public float x() {
-        return x;
-    }
-
-    public float y() {
-        return y;
-    }
-}
