+++ app/src/test/java/io/github/clowngraphics/rerenderer/triangulation/ReadOnlyVector2f.java
@@ -1,0 +1,19 @@
+package io.github.clowngraphics.rerenderer.triangulation;
+
+public class ReadOnlyVector2f implements Vector2f {
+    private final float x;
+    private final float y;
+
+    public ReadOnlyVector2f(float x, float y) {
+        this.x = x;
+        this.y = y;
+    }
+
+    public float x() {
+        return x;
+    }
+
+    public float y() {
+        return y;
+    }
+}
