+++ app/src/test/java/org/example/AppTest.java
@@ -3,12 +3,13 @@
  */
 package org.example;
 
-import io.github.ClownGraphics.Rerenderer.App;
+import io.github.clowngraphics.rerenderer.App;
 import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.*;
 
 class AppTest {
-    @Test void appHasAGreeting() {
+    @Test
+    void appHasAGreeting() {
         App classUnderTest = new App();
         assertNotNull(classUnderTest.getGreeting(), "app should have a greeting");
     }
